Terminate Journeys on encounters having particular values

Helper function for second pass at journey building. This time, validates / Checks that encounters having characteristics that indicate the encounter should terminate its journey

terminate_journeys_by(journeys, ...)

Arguments

journeys
the dataframe representing the journey-keyed encounters

...
Conditions to pass on to check an encounter should terminate its journey

Value

The df, but with the journey key reworked to reflect the termination. Journey length, etc, should also be affected.

TODO

Set this as internal, wrap it in something

See also

Examples

if (FALSE) {
journeys <- encounters %>%
   link_encounters_by_timestamps()

dispositions <- c("Departed - Treatment Completed", "Death Type 1")
rekeyed <- journeys %>%
   terminate_journeys_by(DISCH_DISPOSITION %in% dispositions)
}